+++ src/main/java/com/ktb/joing/common/config/SecurityConfig.java
@@ -50,6 +50,7 @@ public class SecurityConfig {
         http.securityMatcher("/**") // 모든 요청에 대해
                 .authorizeHttpRequests(auth -> auth
                         .requestMatchers("/", "/healthz", "/oauth2/**", "/login/**").permitAll()
+                        .requestMatchers("/api/v1/users/signup/**").hasAuthority("TEMP_USER")
                         .anyRequest().authenticated()
                 );
