--- src/main/java/com/ktb/joing/domain/auth/redis/TempUserRepository.java
@@ -1,6 +1,0 @@
-package com.ktb.joing.domain.auth.redis;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface TempUserRepository extends CrudRepository<TempUser, String> {
-}
