+++ src/main/java/com/ktb/joing/domain/auth/repository/TempUserRepository.java
@@ -1,0 +1,7 @@
+package com.ktb.joing.domain.auth.repository;
+
+import com.ktb.joing.domain.auth.entity.TempUser;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TempUserRepository extends JpaRepository<TempUser, String> {
+}
