+++ src/main/java/com/ktb/joing/domain/user/entity/Role.java
@@ -1,5 +1,5 @@
 package com.ktb.joing.domain.user.entity;
 
 public enum Role {
-    ROLE_USER, ROLE_ADMIN
+    ROLE_TEMP_USER, ROLE_USER, ROLE_ADMIN
 }
